Refactor common types
A lot of types across verbs in ros2 topic are shared. This
PR refactors the types to be used by different verbs preventing
redefinition of the same types in different verbs. There is no
behaviour changed introduced.
